+++ ChangeLog.md
@@ -20,6 +20,9 @@ See docs/process.md for more on how version tagging works.
 
 4.0.14 (in development)
 -----------------------
+- The wasm offset converter was removed along with the `WASM_OFFSET_CONVERTER`
+  setting. This feature only existed to work around an old v8 bug that was fixed
+  back in 2019. (#24963)
 
 4.0.13 - 08/14/25
 -----------------